pyproject.toml
[project]
name = "user-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.8",
    "django-mptt>=0.18.0",
    "djangorestframework-simplejwt>=5.5.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "uvicorn>=0.38.0",
    "uvicorn-worker>=0.4.0",
]

[dependency-groups]
dev = [
    "django-admin>=2.0.2",
    "pytest>=8.4.2",
    "ruff>=0.14.4",
]

[tool.black]
line-length = 120
skip-string-normalization = true
